requirements.txt
Django >= 1.8.0
django-crispy-forms >= 1.6.0
Pillow >= 3.1.1
mistune >= 0.7.1
pygments >= 2.1
django-contrib-comments
django-fluent-comments >= 1.2
django-threadedcomments >= 1.0.1
django_braces >= 1.8.1


